pyproject.toml
[tool.poetry]
name = "ml-flow"
version = "0.1.0"
description = ""
authors = ["Emmanuel Martínez <emartinez@seebak.com.mx>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
aws-cdk-lib = "^2.0.0rc1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
